% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_deck.R
\name{build_deck}
\alias{build_deck}
\title{make a deck of cards with 52 elements and unicode symbols for suits}
\usage{
build_deck()
}
\description{
make a deck of cards with 52 elements and unicode symbols for suits
}
