% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c5000}
\alias{c5000}
\title{read depth for 5000 positions on pasilla chr4, from example(coverage) for GenomicAlignments}
\format{
numeric vector
}
\usage{
c5000
}
\description{
read depth for 5000 positions on pasilla chr4, from example(coverage) for GenomicAlignments
}
\keyword{datasets}
