% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_results}
\alias{de_results}
\title{topTable excerpt for SRP045638, see vGene}
\format{
data.frame
}
\usage{
de_results
}
\description{
topTable excerpt for SRP045638, see vGene
}
\keyword{datasets}
