% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doubsim}
\alias{doubsim}
\title{simulated draws of top card from a shuffled deck of cards}
\format{
character vector
}
\usage{
doubsim
}
\description{
simulated draws of top card from a shuffled deck of cards
}
\keyword{datasets}
