% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doubsim3}
\alias{doubsim3}
\title{simulated draws of top card from a shuffled  deck of cards}
\format{
character vector
}
\usage{
doubsim3
}
\description{
simulated draws of top card from a shuffled  deck of cards
}
\keyword{datasets}
