% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egr1_ex}
\alias{egr1_ex}
\title{RNASeq2GeneNorm measures for EGR1 from TCGA ACC samples}
\format{
numeric vector
}
\usage{
egr1_ex
}
\description{
RNASeq2GeneNorm measures for EGR1 from TCGA ACC samples
}
\keyword{datasets}
