% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_deck.R
\name{faces}
\alias{faces}
\title{get faces of a collection of cards}
\usage{
faces(x)
}
\arguments{
\item{x}{a card from a deck made by `build_deck`}
}
\description{
get faces of a collection of cards
}
\examples{
d = build_deck()
d[1]
faces(d[1:2])
}
