% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fos_ex}
\alias{fos_ex}
\title{RNASeq2GeneNorm measures for FOS from TCGA ACC samples}
\format{
numeric vector
}
\usage{
fos_ex
}
\description{
RNASeq2GeneNorm measures for FOS from TCGA ACC samples
}
\keyword{datasets}
