% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{full_house}
\alias{full_house}
\title{test for full house}
\usage{
full_house(x)
}
\arguments{
\item{x}{a vector of cards}
}
\description{
test for full house
}
\examples{
d = build_deck()
h1 = c(d[1], d[14], d[27], d[2], d[15])
h1
full_house(h1)
h2 = c(d[1], d[14], d[27], d[2], d[16])
h2
full_house(h2)
}
