% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_exc_chr20_b38}
\alias{gtex_exc_chr20_b38}
\title{data.frame with GTEx lung eQTL data and metadata, a selection from chr20}
\format{
data.frame
}
\usage{
gtex_exc_chr20_b38
}
\description{
data.frame with GTEx lung eQTL data and metadata, a selection from chr20
}
\keyword{datasets}
