% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{is_flush}
\alias{is_flush}
\title{test for flush}
\usage{
is_flush(x)
}
\arguments{
\item{x}{a vector of cards}
}
\description{
test for flush
}
\examples{
d = build_deck()
h1 = c(d[1:4], d[6])
h1
is_flush(h1)
}
