% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonlindat}
\alias{nonlindat}
\title{a data frame with y and x in a sigmoidal relationship (with error and a covariate z)}
\format{
data.frame
}
\usage{
nonlindat
}
\description{
a data frame with y and x in a sigmoidal relationship (with error and a covariate z)
}
\examples{
data(nonlindat)
plot(y~x, data=nonlindat, col=z+1, pch=19)
}
\keyword{datasets}
