% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{one_pair}
\alias{one_pair}
\title{test for one pair}
\usage{
one_pair(x)
}
\arguments{
\item{x}{a vector of cards}
}
\value{
logical(1)
}
\description{
test for one pair
}
\examples{
d = build_deck()
hand_1 = c(d[1], d[14], d[2:4] )
hand_1
one_pair(hand[1])
hand_2 = c(d[1], d[6], d[2:4] )
hand_2
one_pair(hand[2])
}
