% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCHCA.R
\name{queryCHCA}
\alias{queryCHCA}
\title{use a snapshot of CuratedQueryAtlasR through duckdb}
\usage{
queryCHCA(con, inst = FALSE)
}
\arguments{
\item{con}{connection to duckdb::duckdb(...)}

\item{inst}{logical(1) if TRUE install httpfs in duckdb}
}
\value{
dplyr::tbl for metadata
}
\description{
use a snapshot of CuratedQueryAtlasR through duckdb
}
\examples{
con = DBI::dbConnect(duckdb::duckdb())
queryCHCA(con) 
}
