% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_deck.R
\name{suits}
\alias{suits}
\title{get suits of a collection of cards}
\usage{
suits(x)
}
\arguments{
\item{x}{a card from a deck made by `build_deck`}
}
\description{
get suits of a collection of cards
}
\examples{
d = build_deck()
d[1]
suits(d[1:2])
}
