% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker.R
\name{two_pairs}
\alias{two_pairs}
\title{test for two pairs}
\usage{
two_pairs(x)
}
\arguments{
\item{x}{a vector of cards}
}
\description{
test for two pairs
}
\examples{
d = build_deck()
h1 = c(d[1], d[14], d[27], d[2], d[15])
h1
two_pairs(h1)
}
