% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vGene}
\alias{vGene}
\title{vGene result for a recount dataset, SRP045638}
\format{
limma EList
}
\usage{
vGene
}
\description{
vGene result for a recount dataset, SRP045638
}
\note{
see https://lcolladotor.github.io/cshl_rstats_genome_scale_2023/differential-gene-expression-analysis-with-limma.html#differential-expression
}
\keyword{datasets}
