% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yy1_ex}
\alias{yy1_ex}
\title{RNASeq2GeneNorm measures for YY1 from TCGA ACC samples}
\format{
numeric vector
}
\usage{
yy1_ex
}
\description{
RNASeq2GeneNorm measures for YY1 from TCGA ACC samples
}
\keyword{datasets}
